[arcane,materials] Renomme 'ComponentCell' en 'ConstituentItem' et le déplace dans son propre fichier.
